+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## Head
+* remove short open tags on functions files to satisfy wpcli
+
 ## 0.8.0 | 10.06.2016
 * update/clean node_modules (update gulp-if version, remove deprecated minify-css, add cssnano)
 * preset image processing quality to 100%