+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
-## Head
+## 0.9.0 | 27.06.2016
+* add onepage template
 * remove short open tags on functions files to satisfy wpcli
 * fix title, close functions
 * cleanups, file re-organizations